promotion是取最优 修改好了 我的锅

+++ src/main/java/logic/hotel/HotelInfo.java
@@ -12,5 +12,5 @@ public interface HotelInfo {
 	
 	public HotelVO getHotelnfo(String hotelId);
 	
-	public int getRoomPrice(String hotelID, RoomType roomType);
+	public double getRoomPrice(String hotelID, RoomType roomType);
 }
